requirements.txt
firebase-admin==6.2.0
pyrebase4==4.7.1
python-dotenv==1.0.0
bcrypt==4.0.1
colorama==0.4.6